# Backend/docs/swagger.yaml
basePath: /
definitions:
  db.Admin:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      password:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  db.LoginAdmin:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  db.LoginUser:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  db.RegisterAdmin:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  db.RegisterUser:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  db.User:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      password:
        type: string
      provider:
        type: string
      provider_id:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:4000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server Petstore server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /admin/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: body data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/db.LoginAdmin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.Admin'
        "400":
          description: Bad Request
      summary: Admin login
      tags:
      - Admin
  /admin/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: body data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/db.RegisterAdmin'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/db.Admin'
        "400":
          description: Bad Request
      summary: Create Admin
      tags:
      - Admin
  /admin/users:
    get:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.Admin'
            type: array
        "400":
          description: Bad Request
      summary: User List
      tags:
      - Admin
  /user/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: body data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/db.LoginUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.User'
        "400":
          description: Bad Request
      summary: User login
      tags:
      - User
  /user/profile:
    get:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.User'
        "400":
          description: Bad Request
      summary: User profile
      tags:
      - User
  /user/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: body data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/db.RegisterUser'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/db.User'
        "400":
          description: Bad Request
      summary: Create User
      tags:
      - User
swagger: "2.0"